+++ LoanPayment/dbo/Tables/Strategy.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [dbo].[Strategy] (
-    [ExtraPerMonth]           FLOAT (53)     NOT NULL,
+    [ExtraPerMonth]           DECIMAL (9, 2) NOT NULL,
     [ExtraPerMonthCalcMethod] NVARCHAR (255) NOT NULL,
     [MonthsDelay]             INT            NOT NULL,
     [SortOrder]               NVARCHAR (255) NOT NULL,
@@ -8,6 +8,8 @@
     [UseSortOrderGroup]       NVARCHAR (255) NOT NULL,
     CONSTRAINT [PK_Strategy] PRIMARY KEY CLUSTERED ([StrategyId] ASC)
 );
+
+
